start refactoring of shared 1.2 / 2.0 functions.  removed 1.2 specific types
