--- packages/main/src/handlers/fdc3/1.2/types/FDC3Data.ts
@@ -1,30 +1,0 @@
-import { Context, IntentMetadata } from 'fdc3-1.2';
-import { DirectoryApp } from '/@/directory/directory';
-
-/**
- * representation of an FDC3 App - whether it is running (connected) or not (directory only)
- */
-export interface FDC3App {
-  type: string;
-  details: FDC3AppDetail;
-  intent?: string; //optional intent associated with the app
-}
-
-/**
- * A collection of apps associated with an intent
- */
-export interface IntentInstance {
-  intent: IntentMetadata;
-  apps: Array<FDC3App>;
-}
-
-export interface FDC3AppDetail {
-  instanceId?: string;
-  title?: string;
-  directoryData?: DirectoryApp;
-}
-
-export interface ResolverDetail {
-  intent?: string;
-  context?: Context;
-}
