+++ packages/main/src/handlers/fdc3/2.0/raiseIntent.ts
@@ -1,274 +1,21 @@
-import {
-  ResolveError,
-  AppIdentifier,
-  IntentResolution,
-  IntentResult,
-  IntentMetadata,
-} from '@finos/fdc3';
+import { ResolveError, AppIdentifier } from '@finos/fdc3';
 import { getRuntime } from '/@/index';
 import { View } from '/@/view';
 import { RuntimeMessage } from '/@/handlers/runtimeMessage';
-import {
-  FDC3App,
-  IntentInstance,
-  FDC3AppDetail,
-} from '/@/handlers/fdc3/1.2/types/FDC3Data';
+import { FDC3App, FDC3AppDetail } from '/@/types/FDC3Data';
 import utils from '/@/utils';
 import { FDC3_2_0_TOPICS } from './topics';
 import { FDC3_1_2_TOPICS } from '/@/handlers/fdc3/1.2/topics';
 import { ipcMain } from 'electron';
-import { RUNTIME_TOPICS } from '/@/handlers/runtime/topics';
+import {
+  buildIntentInstanceTree,
+  sortApps,
+  resolveIntent,
+} from '../lib/raiseIntent';
 import {
   DirectoryApp,
   DirectoryAppLaunchDetailsWeb,
 } from '/@/directory/directory';
-
-export const resolveIntent = async (message: RuntimeMessage) => {
-  const runtime = getRuntime();
-
-  //TODO: autojoin the new app to the channel which the 'open' call is sourced from
-  let resolvedView: View | undefined;
-
-  if (!message.data.selected.instanceId) {
-    const data: DirectoryApp = message.data.selected?.directoryData;
-
-    //launch window
-    const runtime = getRuntime();
-    if (runtime) {
-      const win = runtime.createWorkspace();
-      resolvedView = await win.createView(
-        (data.details as DirectoryAppLaunchDetailsWeb).url,
-        {
-          directoryData: data as DirectoryApp,
-        },
-      );
-
-      //set pending intent and context
-      resolvedView?.setPendingIntent(
-        message.data.intent,
-        message.data.context,
-        message.data.id,
-      );
-    }
-  } else {
-    resolvedView = runtime.getView(message.data.selected?.instanceId);
-    //send new intent
-    if (resolvedView && resolvedView.parent) {
-      if (resolvedView.fdc3Version === '1.2') {
-        resolvedView.content.webContents.send(FDC3_1_2_TOPICS.INTENT, {
-          topic: 'intent',
-          data: { intent: message.data.intent, context: message.data.context },
-        });
-      } else {
-        resolvedView.content.webContents.send(FDC3_2_0_TOPICS.INTENT, {
-          topic: 'intent',
-          data: { intent: message.data.intent, context: message.data.context },
-          source: message.data.id,
-        });
-
-        resolvedView.parent.setSelectedTab(resolvedView.id);
-      }
-    }
-  }
-  //close the resolver
-  const resolver = runtime.getResolver();
-  if (resolver) {
-    resolver.close();
-  }
-
-  const result: IntentResolution = {
-    source: {
-      appId: resolvedView?.directoryData?.appId || '',
-      instanceId: resolvedView?.id,
-    },
-    intent: message.data.intent,
-    version: '2.0',
-    getResult: (): Promise<IntentResult> => {
-      return new Promise((resolve) => {
-        //to do...
-        resolve({
-          type: 'emptyResult',
-        });
-      });
-    },
-  };
-
-  return result;
-};
-
-export const _resolveIntent = (
-  message: RuntimeMessage,
-): Promise<IntentResolution> => {
-  return new Promise((resolve, reject) => {
-    //find the app to route to
-    try {
-      const selected = message.data.selected;
-      const sView = selected?.details?.instanceId
-        ? getRuntime().getView(selected.details.instanceId)
-        : null;
-      const source = message.data.source;
-      const intent = message.data.intent;
-      if (intent) {
-        const listeners = getRuntime().getIntentListeners(intent);
-        //let keys = Object.keys(listeners);
-        let appId: string | undefined = undefined;
-        const id = (sView && sView.id) || undefined;
-        listeners.forEach((listener) => {
-          if (listener.source === id) {
-            appId = listener.source;
-          }
-        });
-
-        if (appId) {
-          console.log('send intent from source', source);
-          const app = getRuntime().getView(appId);
-          if (app && app.content) {
-            if (app.fdc3Version === '1.2') {
-              app.content.webContents.send(FDC3_1_2_TOPICS.INTENT, {
-                topic: 'intent',
-                data: {
-                  intent: message.data.intent,
-                  context: message.data.context,
-                },
-                source: source,
-              });
-            } else {
-              app.content.webContents.send(FDC3_2_0_TOPICS.INTENT, {
-                topic: 'intent',
-                data: {
-                  intent: message.data.intent,
-                  context: message.data.context,
-                },
-                source: source,
-              });
-            }
-
-            if (sView && sView.parent && sView.parent.window) {
-              sView.parent.window.webContents.send(RUNTIME_TOPICS.SELECT_TAB, {
-                viewId: sView.id,
-              });
-              const id = (sView && sView.id) || '';
-
-              resolve({
-                source: {
-                  appId: id,
-                },
-                intent: message.data.intent,
-                version: '2.0',
-                getResult: (): Promise<IntentResult> => {
-                  return new Promise((resolve) => {
-                    //to do...
-                    resolve({
-                      type: 'emptyResult',
-                    });
-                  });
-                },
-              });
-            }
-          }
-
-          //keep array of pending, id by url,  store intent & context, timestamp
-          //when a new window connects, throw out anything more than 2 minutes old, then match on url
-        }
-      }
-    } catch (err) {
-      reject(err);
-    }
-  });
-};
-
-/**
- * create a heirarchy of App Instances grouped by intents
- */
-const buildIntentInstanceTree = (
-  data: Array<FDC3App>,
-): Array<IntentInstance> => {
-  const r: Array<IntentInstance> = [];
-
-  const found: Map<string, Array<FDC3App>> = new Map();
-  const intentMetadata: Array<IntentMetadata> = [];
-
-  if (data) {
-    data.forEach((item) => {
-      const intents =
-        item.details.directoryData?.interop?.intents?.listensFor ?? {};
-
-      Object.keys(intents).forEach((intentId) => {
-        if (!found.has(intentId)) {
-          const intent = intents[intentId];
-          intentMetadata.push({
-            name: intentId,
-            displayName: intent.displayName ?? intentId,
-          });
-          found.set(intentId, [item]);
-        } else {
-          const intents = found.get(intentId);
-          if (intents) {
-            intents.push(item);
-          }
-        }
-      });
-    });
-  }
-
-  intentMetadata.forEach((intent) => {
-    const apps: Array<FDC3App> = found.get(intent.name) || [];
-
-    const entry: IntentInstance = { intent: intent, apps: apps };
-
-    r.push(entry);
-  });
-
-  return r;
-};
-const getAppTitle = (app: FDC3App): string => {
-  const runtime = getRuntime();
-  const view = app.details.instanceId
-    ? runtime.getViews().get(app.details.instanceId)
-    : null;
-  const directory = app.details.directoryData
-    ? app.details.directoryData
-    : null;
-  return directory
-    ? directory.title ?? 'Untitled'
-    : view &&
-      view.content.webContents &&
-      view.content.webContents.hostWebContents
-    ? view.content.webContents.hostWebContents.getTitle()
-    : 'Untitled';
-};
-
-const sortApps = (a: FDC3App, b: FDC3App): number => {
-  //let aTitle = a.details.directoryData ? a.details.directoryData.title : a.details.view.content.webContents.getURL();
-  // let bTitle = b.details.directoryData ? b.details.directoryData.title : b.details.view.content.webContents.getURL();
-  if (a.details) {
-    a.details.title = getAppTitle(a);
-  }
-  if (b.details) {
-    b.details.title = getAppTitle(b);
-  }
-
-  if (
-    a.details &&
-    a.details.title &&
-    b.details &&
-    b.details.title &&
-    a.details.title < b.details.title
-  ) {
-    return -1;
-  }
-  if (
-    a.details &&
-    a.details.title &&
-    b.details &&
-    b.details.title &&
-    a.details.title > b.details.title
-  ) {
-    return 1;
-  } else {
-    return 0;
-  }
-};
 
 export const raiseIntent = async (message: RuntimeMessage) => {
   const runtime = getRuntime();
