+++ packages/main/src/handlers/fdc3/lib/intentListeners.ts
@@ -1,10 +1,16 @@
 import { getRuntime } from '/@/index';
-import { RuntimeMessage } from '/@/handlers/runtimeMessage';
-import { FDC3_2_0_TOPICS } from './topics';
+import { FDC3_1_2_TOPICS } from '../1.2/topics';
+import { FDC3_2_0_TOPICS } from '../2.0/topics';
+import {
+  FDC3Message,
+  ListenerMessageData,
+  IntentListenerData,
+} from '/@/types/FDC3Message';
 
-export const dropIntentListener = async (message: RuntimeMessage) => {
+export const dropIntentListener = async (message: FDC3Message) => {
   const runtime = getRuntime();
-  const id = message.data?.id;
+  const data: ListenerMessageData = message.data as ListenerMessageData;
+  const id = data.listenerId;
   const view = runtime.getView(message.source);
   if (view && id) {
     view.listeners = view.listeners.filter((l) => {
@@ -16,14 +22,15 @@ export const dropIntentListener = async (message: RuntimeMessage) => {
 //wait 2 minutes for pending intents to connect
 const pendingTimeout: number = 2 * 60 * 1000;
 
-export const addIntentListener = async (message: RuntimeMessage) => {
+export const addIntentListener = async (message: FDC3Message) => {
   const runtime = getRuntime();
-  const name = message.data && message.data.intent;
-  const listenerId = message.data && message.data.id;
+  const data: IntentListenerData = message.data as IntentListenerData;
+  const intent = data.intent;
+  const listenerId = data.listenerId;
   const view = runtime.getView(message.source);
 
-  if (name && listenerId && view) {
-    runtime.setIntentListener(name, listenerId, message.source);
+  if (intent && listenerId && view) {
+    runtime.setIntentListener(intent, listenerId, message.source);
 
     //check for pending intents
     const pending_intents = view.getPendingIntents();
@@ -33,12 +40,15 @@ export const addIntentListener = async (message: RuntimeMessage) => {
 
       //next, match on tab and intent
       pending_intents.forEach((pIntent, index) => {
-        if (n - pIntent.ts < pendingTimeout && pIntent.intent === name) {
-          console.log('applying pending intent', pIntent);
+        if (n - pIntent.ts < pendingTimeout && pIntent.intent === intent) {
           //refactor with other instances of this logic
           if (view && view.content) {
-            view.content.webContents.send(FDC3_2_0_TOPICS.INTENT, {
-              topic: 'intent',
+            const topic =
+              view.fdc3Version === '1.2'
+                ? FDC3_1_2_TOPICS.INTENT
+                : FDC3_2_0_TOPICS.INTENT;
+            view.content.webContents.send(topic, {
+              topic: topic,
               data: {
                 intent: pIntent.intent,
                 context: pIntent.context,
@@ -46,6 +56,8 @@ export const addIntentListener = async (message: RuntimeMessage) => {
               source: pIntent.source,
             });
           }
+
+          //remove the applied intent
           view.removePendingIntent(index);
         }
       });
