+++ packages/main/src/types/FDC3Data.ts
@@ -6,6 +6,7 @@ import { DirectoryApp } from '../directory/directory';
 export interface FDC3App {
   type: string;
   details: FDC3AppDetail;
+  intent?: string;
 }
 
 /**
